Fix hang when switching to Notes
In some configurations, the workflow can hang for a very long time (a dozen seconds, or sometimes even indefinitely) when performing an action that opens Notes.

The hang can be skipped by the user simply by manually switching to Notes. This commit explicitely reveals the Notes app whenever it is supposed to be switched to, thus fixing the issue.

As a side effect, this makes the workflow feel more reactive, as the Notes app will instantly be displayed when the workflow is used, even if it then takes a bit of time to focus the actual requested note.