Focus main window before opening results
Before this change, using the workflow to reveal a note would navigate away from the current note, if a single-note window was open and frontmost in Notes. This would even mess with the navigation history of that window.

Now, the workflow makes sure to focus the main Notes window before revealing the requested note.